server/agents/deployment/requirements.txt
fastapi
uvicorn
pydantic
requests
structlog
python-dotenv